pyproject.toml
[project]
name = "my-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.128.0",
    "pymilvus>=2.6.6",
    "pytest>=9.0.2",
    "pyyaml>=6.0.3",
    "tqdm>=4.67.1",
    "uvicorn>=0.40.0",
    "torch",
    "torchvision",
    "transformers>=4.57.6",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
]

[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true  

[tool.uv.sources]
torch = [{ index = "pytorch-cu128" }]
torchvision = [{ index = "pytorch-cu128" }]
